+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2020-03-26
+### Added
+- Function toQuantity(unit)
+- Operators =(equality) and ~(equivalence) for Quantity
+- Implicit conversion from FHIR Quantity to FHIRPath System.Quantity
+
 ## [2.0.0] - 2020-03-11
 ### Changed
 - FHIRPath grammar updated to version 2.0.0 (N1)